+++ R/translate-sql-clause.R
@@ -37,9 +37,8 @@ sql_clause_where <- function(where, con){
 
 sql_clause_limit <- function(limit, con){
   if (!is.null(limit) && !identical(limit, Inf)) {
-    assert_that(is.numeric(limit), length(limit) == 1L, limit >= 0)
     build_sql(
-      "LIMIT ", sql(format(trunc(limit), scientific = FALSE)),
+      "LIMIT ", sql(format(limit, scientific = FALSE)),
       con = con
     )
   }