+++ R/verb-group_by.R
@@ -1,9 +1,15 @@
 # group_by ----------------------------------------------------------------
 
 #' @export
-group_by.tbl_lazy <- function(.data, ..., add = FALSE, .drop = TRUE) {
+#' @importFrom dplyr group_by
+group_by.tbl_lazy <- function(.data, ..., .add = FALSE, add = NULL, .drop = TRUE) {
   dots <- quos(...)
   dots <- partial_eval_dots(dots, vars = op_vars(.data))
+
+  if (!missing(add)) {
+    lifecycle::deprecate_warn("1.0.0", "dplyr::group_by(add = )", "group_by(.add = )")
+    .add <- add
+  }
 
   if (!identical(.drop, TRUE)) {
     stop("`.drop` is not supported with database backends", call. = FALSE)
@@ -13,7 +19,11 @@ group_by.tbl_lazy <- function(.data, ..., add = FALSE, .drop = TRUE) {
     return(.data)
   }
 
-  groups <- group_by_prepare(.data, .dots = dots, add = add)
+  if (".add" %in% names(formals("group_by"))) {
+    groups <- dplyr::group_by_prepare(.data, !!!dots, .add = .add)
+  } else {
+    groups <- dplyr::group_by_prepare(.data, !!!dots, add = .add)
+  }
   names <- purrr::map_chr(groups$groups, as_string)
 
   add_op_single("group_by",
@@ -44,6 +54,7 @@ sql_build.op_group_by <- function(op, con, ...) {
 
 # ungroup -----------------------------------------------------------------
 
+#' @importFrom dplyr ungroup
 #' @export
 ungroup.tbl_lazy <- function(x, ...) {
   add_op_single("ungroup", x)