+++ tests/testthat/helper-src.R
@@ -1,32 +1,33 @@
+on_gha <- function() identical(Sys.getenv("GITHUB_ACTIONS"), "true")
+on_cran <- function() !identical(Sys.getenv("NOT_CRAN"), "true")
+
 if (test_srcs$length() == 0) {
 
-  test_register_src("df", dplyr::src_df(env = new.env(parent = emptyenv())))
+  # test_register_src("df", dplyr::src_df(env = new.env(parent = emptyenv())))
   test_register_con("sqlite", RSQLite::SQLite(), ":memory:")
 
-  if (identical(Sys.getenv("TRAVIS"), "true")) {
-    test_register_con("postgres", RPostgreSQL::PostgreSQL(),
+  if (identical(Sys.getenv("GITHUB_POSTGRES"), "true")) {
+    test_register_con("postgres", RPostgres::Postgres(),
       dbname = "test",
-      user = "travis",
-      password = ""
+      user = "postgres",
+      password = "password"
     )
-  } else {
-    test_register_con("mysql", RMySQL::MySQL(),
-      dbname = "test",
-      host = "localhost",
-      user = Sys.getenv("USER")
-    )
+  } else if (on_gha() || on_cran()) {
+    # Only test with sqlite
+  } else  {
     test_register_con("MariaDB", RMariaDB::MariaDB(),
       dbname = "test",
       host = "localhost",
-      user = Sys.getenv("USER")
+      username = Sys.getenv("USER")
     )
-    test_register_con("postgres", RPostgreSQL::PostgreSQL(),
+    test_register_con("postgres", RPostgres::Postgres(),
       dbname = "test",
       host = "localhost",
       user = ""
     )
   }
 }
+
 
 skip_if_no_db <- function(db) {
   if (!test_srcs$has(db))