+++ tests/testthat/test-verb-compute.R
@@ -58,7 +58,7 @@ test_that("unique index fails if values are duplicated", {
 test_that("compute creates correct column names", {
   out <- memdb_frame(x = 1) %>%
     group_by(x) %>%
-    summarize(n = n()) %>%
+    summarise(n = n()) %>%
     compute() %>%
     collect()
 
@@ -67,13 +67,13 @@ test_that("compute creates correct column names", {
 
 # ops ---------------------------------------------------------------------
 
-test_that("preserved across compute and collapse", {
+test_that("sorting preserved across compute and collapse", {
   df1 <- memdb_frame(x = sample(10)) %>% arrange(x)
 
   df2 <- compute(df1)
-  expect_equal(op_sort(df2), list(~x))
+  expect_equal(get_expr(op_sort(df2)[[1]]), quote(x))
 
   df3 <- collapse(df1)
-  expect_equal(op_sort(df3), list(~x))
+  expect_equal(get_expr(op_sort(df3)[[1]]), quote(x))
 })